travis-ci: add config file for testing firmware handoff on qemu_arm64
The existing 'conf.qemu_arm64_na' uses 'u-boot.bin' as bios, but for
the case of TF-A, 'u-boot.bin' is BL33 and the assembled 'flash.bin'
is the bios instead, which will be built and assembled at run-time
via a post-buildman script and copied to '/tmp'.

Moreover, for running TF-A and OP-TEE on qemu, '-m 4G' and
'virtualization=false,secure=on' are required, and '-semihosting'
should be replaced with its modern alias
'-semihosting-config enable=on,target=native'.

Last, point to the new helper script 'swtpm-tfa'.

Signed-off-by: Raymond Mao <raymond.mao@linaro.org>